phase0-web-interface/requirements.txt
# Phase 0 Web Interface Requirements
# Flask and web framework
Flask==2.3.2
Flask-SocketIO==5.3.4

# YAML support for configuration
PyYAML==6.0

# WebSocket support
python-socketio==5.8.0
python-engineio==4.7.1

# Google Cloud dependencies (for agent integration)
google-cloud-aiplatform==1.26.1
google-cloud-secret-manager==2.16.2
google-auth==2.19.1

# HTTP requests
requests==2.31.0

# Environment variables support
python-dotenv==1.0.0

# Development dependencies
gunicorn==20.1.0  # For production deployment

# Optional: For enhanced logging
colorlog==6.7.0